Function programming in python
As we all know python is a very powerful language, it will allow us to write our code in less time with more optimization. In python, we can do functional programming which will help us to reduce the lines of code and provide us the efficient output.

Few Concepts of Functional Programming:-
1. Higher Order Function.
2. Currying.
3. Immutability.
4. Clousers.

In the above code, I tried to explain the Higher-order function and the Currying.